Criação do teste de retorno paginado e configuração da injeção de dependecias

+++ domain/src/main/java/com/jiankowalski/admin/catalogo/domain/Pagination.java
@@ -6,7 +6,7 @@ import java.util.function.Function;
 public record Pagination<T>(
         int currentPage,
         int perPage,
-        int total,
+        long total,
         List<T> items
 ) {
     public <R> Pagination<R> map(final Function<T, R> mapper) {
